+++ actions/core/install.sh
@@ -17,13 +17,15 @@ core_install ()
         # Clone
         git clone "$CORE_REPO" "$CORE_DIR" | tee -a "$commander_log"
         cd "$CORE_DIR"
-        git checkout master | tee -a "$commander_log"
+        git checkout "$CORE_BRANCH" | tee -a "$commander_log"
 
-        # Install
-        lerna bootstrap | tee -a "$commander_log"
-
+        # Configure
         core_configure
 
+        ascii
+
         success "Installed KAPU Core!"
+
+        core_version
     fi
 }