+++ menus/explorer/help.sh
@@ -8,8 +8,24 @@ menu_manage_explorer_help ()
     text_white "    installs the KAPU blockchain explorer on your server."
     echo
 
+    text_yellow "    Stop KAPU Explorer (K)"
+    text_white "    stops the KAPU Explorer process (shown only if the KAPU Explorer process is running)."
+    echo
+
     text_yellow "    Start KAPU Explorer (S)"
-    text_white "    starts the KAPU blockchain explorer process (only visible if you have KAPU explorer installed)."
+    text_white "    starts the KAPU Explorer process (shown only if KAPU Explorer is installed)."
+    echo
+
+    text_yellow "    Restart KAPU Explorer (R)"
+    text_white "    restarts (stops and then starts) the KAPU Explorer process (shown only if KAPU Explorer is installed and running)."
+    echo
+
+    text_yellow "    Update KAPU Explorer (U)"
+    text_white "    updates KAPU Explorer if a new version is available (shown only if KAPU Explorer is installed)."
+    echo
+
+    text_yellow "    Uninstall KAPU Explorer (P)"
+    text_white "    uninstalls KAPU Explorer from your system (shown only if KAPU Explorer is installed)."
     echo
 
     text_yellow "    Show Log (L)"
@@ -21,7 +37,7 @@ menu_manage_explorer_help ()
 
     divider
 
-    text_blue "    For more information head over to https://arkdocs.readme.io/"
+    text_blue "    For more information head over to https://docs.ark.io/"
     echo
 
     press_to_continue