+++ menus/explorer/menu.sh
@@ -7,13 +7,13 @@ menu_manage_explorer ()
     explorer_status
 
     if [[ "$STATUS_EXPLORER" = "On" ]]; then
-        text_white "    P. Uninstall KAPU Explorer"
         text_white "    K. Stop KAPU Explorer"
         text_white "    R. Restart KAPU Explorer"
-        text_white "    U. Update KAPU Explorer"
     else
         if [[ -d "$EXPLORER_DIR" ]]; then
             text_white "    S. Start KAPU Explorer"
+            text_white "    U. Update KAPU Explorer"
+            text_white "    P. Uninstall KAPU Explorer"
         else
             text_white "    I. Install KAPU Explorer"
         fi