+++ menus/main/menu.sh
@@ -5,33 +5,27 @@ menu_main ()
     refresh_state
 
     if [[ "$STATUS_RELAY" = "On" ]]; then
-        local relay="Relay: $(text_green ${STATUS_RELAY})"
+        local relay="$(text_white "Relay:")  $(text_green ${STATUS_RELAY})"
     else
-        local relay="Relay: $(text_red ${STATUS_RELAY})"
+        local relay="$(text_white "Relay:") $(text_red ${STATUS_RELAY})"
     fi
 
     if [[ "$STATUS_FORGER" = "On" ]]; then
-        local forger="Forger: $(text_green ${STATUS_FORGER})"
+        local forger="$(text_white "Forger:")  $(text_green ${STATUS_FORGER})"
     else
-        local forger="Forger: $(text_red ${STATUS_FORGER})"
+        local forger="$(text_white "Forger:") $(text_red ${STATUS_FORGER})"
     fi
 
     if [[ "$STATUS_NTP" = "On" ]]; then
-        local ntp="NTP: $(text_green ${STATUS_NTP})"
+        local ntp="$(text_white "NTP:")  $(text_green ${STATUS_NTP})"
     else
-        local ntp="NTP: $(text_red ${STATUS_NTP})"
+        local ntp="$(text_white "NTP:") $(text_red ${STATUS_NTP})"
     fi
 
     if [[ "$STATUS_PGSQL" = "On" ]]; then
-        local pgsql="PG: $(text_green ${STATUS_PGSQL})"
+        local pgsql="$(text_white "PG:")  $(text_green ${STATUS_PGSQL})"
     else
-        local pgsql="PG: $(text_red ${STATUS_PGSQL})"
-    fi
-
-    if [[ "$STATUS_REDIS" = "On" ]]; then
-        local redis="Redis: $(text_green ${STATUS_REDIS})"
-    else
-        local redis="Redis: $(text_red ${STATUS_REDIS})"
+        local pgsql="$(text_white "PG:") $(text_red ${STATUS_PGSQL})"
     fi
 
     if [[ -d "$CORE_DIR" ]]; then
@@ -42,16 +36,34 @@ menu_main ()
         fi
     fi
 
+    local core_version_raw="Core: ${VERSION_CORE}"
+    local node_version_raw="Node: ${VERSION_NODE}"
+    local pg_version_raw="PG: ${VERSION_PSQL}"
+
+    local core_version="$(text_white "Core:") $(text_blue ${VERSION_CORE})"
+    local node_version="$(text_white "NodeJS:") $(text_blue ${VERSION_NODE})"
+    local pg_version="$(text_white "PG:") $(text_blue ${VERSION_PSQL})"
+
+    local pad=$(printf '%0.1s' " "{1..30})
+    local whitespace=$((61 - ${#core_version_raw} - ${#node_version_raw} - ${#pg_version_raw}))
+    local padding=$(($whitespace / 2))
+
     trap - INT
 
     ascii
 
-    text_white "    Core: $(text_blue ${VERSION_CORE})    NodeJS: $(text_blue ${VERSION_NODE})    PG: $(text_blue ${VERSION_PSQL})    Redis: $(text_blue ${VERSION_REDIS})"
+    printf '    %s' "$core_version"
+    printf '%*.*s' 0 $(($padding)) "$pad"
+    printf '%s' "$node_version"
+    if [ $((whitespace % 2)) -eq 1 ]; then
+        padding=$(($padding + 1))
+    fi
+    printf '%*.*s' 0 $(($padding)) "$pad"
+    printf '%s\n' "$pg_version"
 
     divider
 
-    text_white "    $relay    $forger    $ntp    $pgsql    $redis"
-
+    echo "    $relay         $forger         $ntp         $pgsql"
 
     divider_double