+++ modules/state.sh
@@ -2,11 +2,8 @@
 
 refresh_state ()
 {
-    get_versions
-
     ntp_status
     pgsql_status
-    redis_status
     relay_status
     forger_status
     explorer_status