+++ CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Changed
+- Moved all app configuration to `ecosystem.config.js` [See also](https://pm2.io/doc/en/runtime/guide/ecosystem-file/)
+- A new npm dependency `dotenv` has been added. When updating from a previous  
+version it is necessary to install it by executing `sudo npm install -g dotenv`  
+followed by `npm link dotenv` inside the commander directory.
+- When using `bip38` the password is no longer exposed in the process monitor.
+- Starting the relay/forger/explorer manually is now easier:
+    Relay:
+    `pm2 start --only ark-core-relay`
+    Forger without bip38 encryption:
+    `pm2 start --only ark-core-forger`
+    Forger with bip38 encryption:
+    `pm2 start --only ark-core-forger -- --password ark`
+    Explorer:
+    `pm2 start --only ark-core-explorer`
+
+### Removed
+- Redis
+
 ## 0.2.0 - 2018-06-19
 ### Added
 - Log level configuration via `.env`